requirements.txt
django==2.1.3
djangorestframework
django-cors-headers
mysqlclient==1.3.12
gunicorn
